Fixed comment handling in emacs 21 and later. Comments weren't being fontified or parsed at all.
Eliminated the function dylan-install-font-lock-defaults, which was
apparently a hack to manually enable dylan-mode on emacs 21 and later
by calling it from .emacs instead of actually checking the emacs
version at runtime (which I changed the code to do). Be sure to remove
any calls to this function from your .emacs file if you were using it.

Replaced all use of make-variable-buffer-local with
make-local-variable. emacs documentation says not to use the former,
generally.

Changed one occurrence of "(list ...)" to "'(...)". I'm not entirely
sure why, but the original code causes a crash/error in emacs 20.

Changed the mode name to a capitalized "Dylan" instead of "dylan".