Include symbol characters in font-lock regex
Makes it so that let foo-bar highlights foo-bar as a variable name
rather than just foo.

This change also changes the face for slot names to always be
font-lock-variable-name-face; there was a conflict with
another font lock rule (variable :: type) where if the slot had
a type, it would use -variable-name-face, but if not it would use
-function-name-face.

Fixes #9.

Signed-off-by: Erik Charlebois <erikcharlebois@gmail.com>